+++ lib/screens/scan_page.dart
